+++ hs_restclient/endpoints/resources.py
@@ -1,3 +1,5 @@
+from ..generators import resultsListGenerator
+
 class BaseEndpoint(object):
     def __init__(self, hs):
         self.hs = hs
@@ -79,6 +81,7 @@ class FunctionsSubEndpoint(object):
         r = self.hs._request('POST', url, None, payload)
         return r
 
+
 class ResourceEndpoint(BaseEndpoint):
     def __init__(self, hs, pid):
         super(ResourceEndpoint, self).__init__(hs)
@@ -140,3 +143,77 @@ class ResourceEndpoint(BaseEndpoint):
                                                           pid=self.pid)
         r = self.hs._request('POST', url)
         return r
+
+
+class ResourceList(BaseEndpoint):
+    def __init__(self, hs, **kwargs):
+        super(ResourceList, self).__init__(hs)
+
+        """
+        Query the GET /hsapi/resource/ REST end point of the HydroShare server.
+
+        :param creator: Filter results by the HydroShare user name of resource creators
+        :param owner: Filter results by the HydroShare user name of resource owners
+        :param user: Filter results by the HydroShare user name of resource users (i.e. owner, editor, viewer, public
+            resource)
+        :param group: Filter results by the HydroShare group name associated with resources
+        :param from_date: Filter results to those created after from_date.  Must be datetime.date.
+        :param to_date: Filter results to those created before to_date.  Must be datetime.date.  Because dates have
+            no time information, you must specify date+1 day to get results for date (e.g. use 2015-05-06 to get
+            resources created up to and including 2015-05-05)
+        :param types: Filter results to particular HydroShare resource types.  Must be a sequence type
+            (e.g. list, tuple, etc.), but not a string.
+
+        :raises: HydroShareHTTPException to signal an HTTP error
+        :raises: HydroShareArgumentException for any invalid arguments
+
+        :return: A generator that can be used to fetch dict objects, each dict representing
+            the JSON object representation of the resource returned by the REST end point.  For example:
+
+        >>> for resource in hs.getResourceList():
+        >>>>    print resource
+         {u'bag_url': u'http://www.hydroshare.org/static/media/bags/e62a438bec384087b6c00ddcd1b6475a.zip',
+          u'creator': u'B Miles',
+          u'date_created': u'05-05-2015',
+          u'date_last_updated': u'05-05-2015',
+          u'resource_id': u'e62a438bec384087b6c00ddcd1b6475a',
+          u'resource_title': u'My sample DEM',
+          u'resource_type': u'RasterResource',
+          u'science_metadata_url': u'http://www.hydroshare.org/hsapi/scimeta/e62a438bec384087b6c00ddcd1b6475a/',
+          u'public': True}
+         {u'bag_url': u'http://www.hydroshare.org/static/media/bags/hr3hy35y5ht4y54hhthrtg43w.zip',
+          u'creator': u'B Miles',
+          u'date_created': u'01-02-2015',
+          u'date_last_updated': u'05-13-2015',
+          u'resource_id': u'hr3hy35y5ht4y54hhthrtg43w',
+          u'resource_title': u'Other raster',
+          u'resource_type': u'RasterResource',
+          u'science_metadata_url': u'http://www.hydroshare.org/hsapi/scimeta/hr3hy35y5ht4y54hhthrtg43w/',
+          u'public': True}
+
+
+          Filtering (have):
+
+          /hsapi/resourceList/?from_date=2015-05-03&to_date=2015-05-06
+          /hsapi/resourceList/?user=admin
+          /hsapi/resourceList/?owner=admin
+          /hsapi/resourceList/?creator=admin
+          /hsapi/resourceList/?group=groupname
+          /hsapi/resourceList/?types=GenericResource&types=RasterResource
+
+          Filtering (need):
+
+          /hsapi/resourceList/?sharedWith=user
+
+        """
+        url = "{url_base}/resource/".format(url_base=self.hs.url_base)
+
+        params = kwargs
+        if 'from_date' in kwargs:
+            params['from_date'] = kwargs['from_date'].strftime('%Y-%m-%d')
+        if 'to_date' in kwargs:
+            params['to_date'] = kwargs['to_date'].strftime('%Y-%m-%d')
+        if 'types' in kwargs:
+            params['type'] = kwargs.pop('types')
+
+        self.list = resultsListGenerator(self.hs, url, params)
