+++ tests/mocks/hydroshare.py
@@ -408,3 +408,27 @@ def resourceUploadFile_post(url, request):
         # 404.
         return response(404, {}, HEADERS, None, 5, request)
     return response(200, content, HEADERS, None, 5, request)
+
+
+@urlmatch(netloc=NETLOC, method=GET)
+def resourcesListByKeyword_get(url, request):
+    file_path = url.netloc + url.path + 'resourceList-keyword'
+    try:
+        content = Resource(file_path).get()
+    except EnvironmentError:
+        # catch any environment errors (i.e. file does not exist) and return a
+        # 404.
+        return response(404, {}, HEADERS, None, 5, request)
+    return response(200, content, HEADERS, None, 5, request)
+
+
+@urlmatch(netloc=NETLOC, method=GET)
+def resourcesListByBoundingBox_get(url, request):
+    file_path = url.netloc + url.path + 'resourceList-bounding-box'
+    try:
+        content = Resource(file_path).get()
+    except EnvironmentError:
+        # catch any environment errors (i.e. file does not exist) and return a
+        # 404.
+        return response(404, {}, HEADERS, None, 5, request)
+    return response(200, content, HEADERS, None, 5, request)
