+++ tests/www.hydroshare.org/hsapi/resource/resourceList-keyword
@@ -1,0 +1,37 @@
+{
+    "count": 2,
+    "next": null,
+    "previous": null,
+    "results": [
+        {
+            "resource_type": "GenericResource",
+            "resource_title": "Untitled resource 2",
+            "resource_id": "1a63d6c24b9b4b42b9a97abfab94f9c6",
+            "creator": "bmiles",
+            "date_created": "05-19-2015",
+            "date_last_updated": "05-19-2015",
+            "public": false,
+            "discoverable": false,
+            "shareable": false,
+            "immutable": false,
+            "published": false,
+            "bag_url": "http://www.hydroshare.org/static/media/bags/1a63d6c24b9b4b42b9a97abfab94f9c6.zip",
+            "science_metadata_url": "http://www.hydroshare.org/hsapi/scimeta/1a63d6c24b9b4b42b9a97abfab94f9c6/"
+        },
+        {
+            "resource_type": "RasterResource",
+            "resource_title": "Create via REST API and Python client library 3",
+            "resource_id": "408acb6b870d4297b8b036edcd9c3c58",
+            "creator": "bmiles",
+            "date_created": "05-22-2015",
+            "date_last_updated": "05-22-2015",
+            "public": true,
+            "discoverable": true,
+            "shareable": true,
+            "immutable": false,
+            "published": false,
+            "bag_url": "http://www.hydroshare.org/static/media/bags/408acb6b870d4297b8b036edcd9c3c58.zip",
+            "science_metadata_url": "http://www.hydroshare.org/hsapi/scimeta/408acb6b870d4297b8b036edcd9c3c58/"
+        }
+    ]
+}
