+++ tests/www.hydroshare.org/resourceList-bounding-box
@@ -1,0 +1,52 @@
+{
+    "count": 5,
+    "next": "http://www.hydroshare.org/hsapi/resource/?page=2",
+    "previous": null,
+    "results": [
+        {
+            "resource_type": "GenericResource",
+            "resource_title": "Untitled resource 1",
+            "resource_id": "c165b6f8ec64405da004bfb8890c35ae",
+            "creator": "selimnairb",
+            "date_created": "05-18-2015",
+            "date_last_updated": "05-18-2015",
+            "public": false,
+            "discoverable": false,
+            "shareable": false,
+            "immutable": false,
+            "published": false,
+            "bag_url": "http://www.hydroshare.org/static/media/bags/c165b6f8ec64405da004bfb8890c35ae.zip",
+            "science_metadata_url": "http://www.hydroshare.org/hsapi/scimeta/c165b6f8ec64405da004bfb8890c35ae/"
+        },
+        {
+            "resource_type": "GenericResource",
+            "resource_title": "Untitled resource 2",
+            "resource_id": "1a63d6c24b9b4b42b9a97abfab94f9c6",
+            "creator": "bmiles",
+            "date_created": "05-19-2015",
+            "date_last_updated": "05-19-2015",
+            "public": false,
+            "discoverable": false,
+            "shareable": false,
+            "immutable": false,
+            "published": false,
+            "bag_url": "http://www.hydroshare.org/static/media/bags/1a63d6c24b9b4b42b9a97abfab94f9c6.zip",
+            "science_metadata_url": "http://www.hydroshare.org/hsapi/scimeta/1a63d6c24b9b4b42b9a97abfab94f9c6/"
+        },
+        {
+            "resource_type": "RasterResource",
+            "resource_title": "Create via REST API and Python client library",
+            "resource_id": "5a6e5a992b1046ae820fd3e79ee36107",
+            "creator": "selimnairb",
+            "date_created": "05-20-2015",
+            "date_last_updated": "05-20-2015",
+            "public": true,
+            "discoverable": true,
+            "shareable": true,
+            "immutable": false,
+            "published": false,
+            "bag_url": "http://www.hydroshare.org/static/media/bags/5a6e5a992b1046ae820fd3e79ee36107.zip",
+            "science_metadata_url": "http://www.hydroshare.org/hsapi/scimeta/5a6e5a992b1046ae820fd3e79ee36107/"
+        }
+    ]
+}
